+++ js/data.js
@@ -1,0 +1,49 @@
+// Константы:
+
+export const OBJECTS_COUNT = 10;
+export const OFFER_TITLE = 'Nice apartments';
+export const MIN_PRICE = 100;
+export const MAX_PRICE = 10000;
+export const MIN_ROOMS = 1;
+export const MAX_ROOMS = 5;
+export const MIN_GUESTS = 1;
+export const MAX_GUESTS = 10;
+export const MIN_LAT = 35.65000;
+export const MAX_LAT = 35.70000;
+export const MIN_LNG = 139.70000;
+export const MAX_LNG = 139.80000;
+export const DESCRIPTION = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
+
+
+export const OFFER_TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+]
+
+export let avatarsList = [
+  'img/avatars/user01.png',
+  'img/avatars/user02.png',
+  'img/avatars/user03.png',
+  'img/avatars/user04.png',
+  'img/avatars/user05.png',
+  'img/avatars/user06.png',
+  'img/avatars/user07.png',
+  'img/avatars/user08.png',
+  'img/avatars/user09.png',
+  'img/avatars/user10.png',
+]
+
+export const PHOTOS_LIST = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+]
+
+export const CHECKIN_OPTIONS = ['12:00', '13:00', '14:00'];
+
+export const CHECKOUT_OPTIONS = ['12:00', '13:00', '14:00'];
+
+export const FEATURES_LIST = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
